Use UUIDs for PipelineIds
This change is being made in Servo for security purposes;
this commit updates the definition to match for Webrender.
See the Servo repo for the reasoning behind this change.

Additionally, remove the TranslatePointToLayerSpace message
from the API; as of servo/servo#11537 this is no longer used.
Servo now uses a unified code path for mouse events whether
WebRender is in use or not. (Mouse events are always dispatched
to the root pipeline, removing the need to translate into
layer space.